+++ application/config/autoload.php
@@ -133,4 +133,4 @@ $autoload['language'] = array();
 |	$autoload['model'] = array('first_model' => 'first');
 */
 
-$autoload['model'] = array('parts', 'robots');
+$autoload['model'] = array('parts', 'robots', 'history');