+++ changelog.md
@@ -8,7 +8,7 @@ ex: JF: added robot images to the assets folder
 
 ## [Unreleased]
 ### Added
-
+- AR: added history view and controller
 - MG: added parts model, parts view, and parts controller
 - AR: added history model
 - AZ: Added a Robots.php model in the models folder. Contains each robot model ID, the part ID's that make up each robot as well as the robots type. Closes #13
@@ -18,6 +18,7 @@ ex: JF: added robot images to the assets folder
 
 
 ### Changed
+- AR: added comments to menubar and template views
 - AR: update history model to include all types in every sub array
 - MG: Updated template to provide some code cleanup.
 - JF: Updated logo